Re-arrange components
