--- client/components/Footer.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class Footer extends Component {
-	render() {
-		return (
-			<footer>
-				<p>WeConnect &copy; 2018</p>
-			</footer>
-		);
-	}
-}
-
-export default Footer;
