--- client/components/Home/HeroImage.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const HeroImage = () => (
-	<div className="hero-image">
-		<h1 className="container catchphrase">Connecting People, Connecting Businesses</h1>
-	</div>
-);
-
-export default HeroImage;
