+++ client/components/common/Footer.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Footer = () => (
+		<footer>
+			<p>WeConnect &copy; 2018</p>
+		</footer>
+);
+
+export default Footer;
