+++ client/components/forms/EditReviewForm.jsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import StarRatingComponent from 'react-star-rating-component';
+
+const EditReviewForm = ({
+  onReviewChange, submitEditedReview, onEditStarClick, formDetails
+}) => (
+  <div>
+    <form onSubmit={submitEditedReview}>
+      <div>
+        <textarea
+          placeholder="Write your review here!"
+          className="form-control"
+          value={formDetails.editedReview}
+          onChange={onReviewChange}
+          name='editedReview'
+          rows="3"
+        >
+        </textarea>
+      </div>
+      <div style={{ fontSize: 25 }}>
+        <StarRatingComponent
+          name='edit'
+          starCount={5}
+          value={formDetails.editedStarRating}
+          onStarClick={onEditStarClick}
+          starColor='#fd654d'
+        />
+      </div>
+      <button className="btn btn-primary mr-2" type="submit">
+        Edit Review
+      </button>
+    </form>
+  </div>
+);
+
+EditReviewForm.propTypes = {
+  review: PropTypes.object.isRequired,
+  star: PropTypes.number,
+  businessId: PropTypes.number.isRequired,
+  reviewId: PropTypes.number.isRequired,
+  submitEditedReview: PropTypes.func.isRequired,
+  onReviewChange: PropTypes.func.isRequired,
+  onEditStarClick: PropTypes.func.isRequired,
+  formDetails: PropTypes.object
+};
+
+export default EditReviewForm;
