+++ package.json
@@ -40,7 +40,6 @@
     "express": "^4.16.2",
     "express-validator": "^5.0.3",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "firebase": "^4.13.1",
     "jsonwebtoken": "^8.2.0",
     "lodash": "^4.17.5",
     "moment": "^2.22.1",
@@ -75,7 +74,7 @@
     "collectCoverage": true,
     "testEnvironment": "node",
     "setupFiles": [
-      "<rootDir>/testSetup.js"
+      "./client/__test__/mocks/localStorageMock.js"
     ]
   },
   "devDependencies": {
